+++ src/SermepaException.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * @file
+ * Implementation file for SermepaException class.
+ */
+
+/**
+ * Type naming the ErrorException class
+ */
+class SermepaException extends ErrorException {
+  // Left blank intentionally.
+}