+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 Sermepa - master Oct 2013
 =========================
+- Improved exception treatment.
 - Added CHANGELOG.md, support to initialize the instance and updated the demo.
 - Initial commit.